+++ src/commands/profiles/add.js
@@ -13,13 +13,13 @@ const FRIENDLY_STORAGE_LOCATIONS = {
   [STORAGE_LOCATIONS.LIBSECRET]: 'using libsecret'
 };
 
-class ProjectsAdd extends BaseCommand {
+class ProfilesAdd extends BaseCommand {
   constructor(argv, config, secureStorage) {
     super(argv, config, secureStorage);
 
     this.accountSid = undefined;
     this.authToken = undefined;
-    this.projectId = undefined;
+    this.profileId = undefined;
     this.force = false;
     this.questions = [];
   }
@@ -28,9 +28,9 @@ class ProjectsAdd extends BaseCommand {
     await super.run();
 
     this.loadArguments();
-    await this.promptForProjectId();
+    await this.promptForProfileId();
 
-    if (!(await this.confirmProjectAndEnvVars()) || !(await this.confirmOverwrite())) {
+    if (!(await this.confirmProfileAndEnvVars()) || !(await this.confirmOverwrite())) {
       this.cancel();
     }
 
@@ -40,7 +40,7 @@ class ProjectsAdd extends BaseCommand {
 
     if (await this.validateCredentials()) {
       await this.saveCredentials();
-      this.logger.info(`Saved ${this.projectId}.`);
+      this.logger.info(`Saved ${this.profileId}.`);
     } else {
       this.cancel();
     }
@@ -49,32 +49,21 @@ class ProjectsAdd extends BaseCommand {
   loadArguments() {
     this.accountSid = this.args['account-sid'];
     this.authToken = this.flags['auth-token'];
-    this.projectId = this.flags.project;
+    this.profileId = this.flags.profile;
     this.force = this.flags.force;
     this.region = this.flags.region;
   }
 
-  async promptForProjectId() {
-    if (!this.projectId) {
-      let counter = 0;
+  async promptForProfileId() {
+    if (!this.profileId) {
       const answer = await this.inquirer.prompt([
         {
-          name: 'projectId',
-          message: this.getPromptMessage(ProjectsAdd.flags.project.description),
-          validate: function (value) {
-            if (!value && counter < 1) {
-              counter++;
-              return 'Shorthand identifier for your Twilio project is required';
-            }
-            return true;
-          }
+          name: 'profileId',
+          message: this.getPromptMessage(ProfilesAdd.flags.profile.description),
+          validate: input => Boolean(input)
         }
       ]);
-      if (!answer.projectId) {
-        this.logger.error('Shorthand identifier for your Twilio project is required');
-        return this.exit(1);
-      }
-      this.projectId = answer.projectId;
+      this.profileId = answer.profileId;
     }
   }
 
@@ -82,7 +71,7 @@ class ProjectsAdd extends BaseCommand {
     if (!this.accountSid) {
       this.questions.push({
         name: 'accountSid',
-        message: this.getPromptMessage(ProjectsAdd.args[0].description),
+        message: this.getPromptMessage(ProfilesAdd.args[0].description),
         validate: input => Boolean(input)
       });
     }
@@ -93,7 +82,7 @@ class ProjectsAdd extends BaseCommand {
       this.questions.push({
         type: 'password',
         name: 'authToken',
-        message: this.getPromptMessage(ProjectsAdd.flags['auth-token'].description),
+        message: this.getPromptMessage(ProfilesAdd.flags['auth-token'].description),
         validate: input => Boolean(input)
       });
     }
@@ -116,14 +105,14 @@ class ProjectsAdd extends BaseCommand {
 
   async confirmOverwrite() {
     let overwrite = true;
-    if (this.userConfig.getProjectById(this.projectId)) {
+    if (this.userConfig.getProfileById(this.profileId)) {
       overwrite = this.force;
       if (!overwrite) {
         const confirm = await this.inquirer.prompt([
           {
             type: 'confirm',
             name: 'overwrite',
-            message: `Overwrite existing project credentials for "${this.projectId}"?`,
+            message: `Overwrite existing profile credentials for "${this.profileId}"?`,
             default: false
           }
         ]);
@@ -133,17 +122,17 @@ class ProjectsAdd extends BaseCommand {
     return overwrite;
   }
 
-  async confirmProjectAndEnvVars() {
+  async confirmProfileAndEnvVars() {
     let affirmative = true;
-    if (this.userConfig.getProjectFromEnvironment()) {
+    if (this.userConfig.getProfileFromEnvironment()) {
       const confirm = await this.inquirer.prompt([
         {
           type: 'confirm',
           name: 'affirmative',
           message:
             'Account credentials are currently stored in environment variables and will take precedence over ' +
-            `the "${this.projectId}" project when connecting to Twilio, unless the "${this.projectId}" project is ` +
-            `explicitly specified. Continue setting up "${this.projectId}" project?`,
+            `the "${this.profileId}" profile when connecting to Twilio, unless the "${this.profileId}" profile is ` +
+            `explicitly specified. Continue setting up "${this.profileId}" profile?`,
           default: false
         }
       ]);
@@ -194,8 +183,8 @@ class ProjectsAdd extends BaseCommand {
       this.exit(1);
     }
 
-    this.userConfig.addProject(this.projectId, this.accountSid, this.region);
-    await this.secureStorage.saveCredentials(this.projectId, apiKey.sid, apiKey.secret);
+    this.userConfig.addProfile(this.profileId, this.accountSid, this.region);
+    await this.secureStorage.saveCredentials(this.profileId, apiKey.sid, apiKey.secret);
     const configSavedMessage = await this.configFile.save(this.userConfig);
 
     this.logger.info(
@@ -207,17 +196,17 @@ class ProjectsAdd extends BaseCommand {
   }
 }
 
-ProjectsAdd.aliases = ['login'];
-ProjectsAdd.description = 'add credentials for an existing Twilio project';
+ProfilesAdd.aliases = ['login'];
+ProfilesAdd.description = 'add a new profile to store Twilio Project credentials and configuration';
 
-ProjectsAdd.flags = Object.assign(
+ProfilesAdd.flags = Object.assign(
   {
     'auth-token': flags.string({
-      description: 'Your Twilio Auth Token for your Twilio project.'
+      description: 'Your Twilio Auth Token for your Twilio Project.'
     }),
     force: flags.boolean({
       char: 'f',
-      description: 'Force overwriting existing project credentials.'
+      description: 'Force overwriting existing profile credentials.'
     }),
     region: flags.string({
       hidden: true
@@ -226,11 +215,11 @@ ProjectsAdd.flags = Object.assign(
   TwilioClientCommand.flags // Yes! We _do_ want the same flags as TwilioClientCommand
 );
 
-ProjectsAdd.args = [
+ProfilesAdd.args = [
   {
     name: 'account-sid',
-    description: 'The Account SID for your Twilio project.'
+    description: 'The Account SID for your Twilio Project.'
   }
 ];
 
-module.exports = ProjectsAdd;
+module.exports = ProfilesAdd;