+++ test/commands/profiles/list.test.js
@@ -1,31 +1,31 @@
 const { expect, test, constants } = require('@twilio/cli-test');
 const { Config, ConfigData } = require('@twilio/cli-core').services.config;
-const ProjectsList = require('../../../src/commands/projects/list');
+const ProfilesList = require('../../../src/commands/profiles/list');
 
 describe('commands', () => {
-  describe('projects', () => {
+  describe('profiles', () => {
     describe('list', () => {
       test
         .twilioCliEnv(Config)
         .stdout()
         .stderr()
-        .twilioCommand(ProjectsList, [])
-        .it('runs projects:list with no projects', ctx => {
+        .twilioCommand(ProfilesList, [])
+        .it('runs profiles:list with no profiles', ctx => {
           expect(ctx.stdout).to.equal('');
-          expect(ctx.stderr).to.contain('No projects have been configured');
+          expect(ctx.stderr).to.contain('No profiles have been configured');
         });
 
       test
         .do(ctx => {
           ctx.userConfig = new ConfigData();
-          ctx.userConfig.addProject('project1', constants.FAKE_ACCOUNT_SID);
+          ctx.userConfig.addProfile('profile1', constants.FAKE_ACCOUNT_SID);
         })
         .twilioCliEnv(Config)
         .stdout()
         .stderr()
-        .twilioCommand(ProjectsList, [])
-        .it('runs projects:list with 1 project', ctx => {
-          expect(ctx.stdout).to.contain('project1');
+        .twilioCommand(ProfilesList, [])
+        .it('runs profiles:list with 1 profile', ctx => {
+          expect(ctx.stdout).to.contain('profile1');
           expect(ctx.stdout).to.contain(constants.FAKE_ACCOUNT_SID);
           expect(ctx.stdout).to.not.contain('Region');
           expect(ctx.stdout.match(/true/g)).to.have.length(1);
@@ -35,16 +35,16 @@ describe('commands', () => {
       test
         .do(ctx => {
           ctx.userConfig = new ConfigData();
-          ctx.userConfig.addProject('project1', constants.FAKE_ACCOUNT_SID);
-          ctx.userConfig.addProject('project2', constants.FAKE_ACCOUNT_SID);
+          ctx.userConfig.addProfile('profile1', constants.FAKE_ACCOUNT_SID);
+          ctx.userConfig.addProfile('profile2', constants.FAKE_ACCOUNT_SID);
         })
         .twilioCliEnv(Config)
         .stdout()
         .stderr()
-        .twilioCommand(ProjectsList, [])
-        .it('runs projects:list with multiple projects', ctx => {
-          expect(ctx.stdout).to.contain('project1');
-          expect(ctx.stdout).to.contain('project2');
+        .twilioCommand(ProfilesList, [])
+        .it('runs profiles:list with multiple profiles', ctx => {
+          expect(ctx.stdout).to.contain('profile1');
+          expect(ctx.stdout).to.contain('profile2');
           expect(ctx.stdout).to.contain(constants.FAKE_ACCOUNT_SID);
           expect(ctx.stdout).to.not.contain('Region');
           expect(ctx.stdout.match(/true/g)).to.have.length(1);
@@ -54,34 +54,34 @@ describe('commands', () => {
       test
         .do(ctx => {
           ctx.userConfig = new ConfigData();
-          ctx.userConfig.addProject('project1', constants.FAKE_ACCOUNT_SID);
-          ctx.userConfig.addProject('project2', constants.FAKE_ACCOUNT_SID);
-          ctx.userConfig.activeProject = 'project1';
+          ctx.userConfig.addProfile('profile1', constants.FAKE_ACCOUNT_SID);
+          ctx.userConfig.addProfile('profile2', constants.FAKE_ACCOUNT_SID);
+          ctx.userConfig.activeProfile = 'profile1';
         })
         .twilioCliEnv(Config)
         .stdout()
         .stderr()
-        .twilioCommand(ProjectsList, [])
-        .it('when the active project is set', ctx => {
-          expect(ctx.stdout).to.contain('project1');
-          expect(ctx.stdout).to.contain('project2');
+        .twilioCommand(ProfilesList, [])
+        .it('when the active profile is set', ctx => {
+          expect(ctx.stdout).to.contain('profile1');
+          expect(ctx.stdout).to.contain('profile2');
           expect(ctx.stdout).to.contain(constants.FAKE_ACCOUNT_SID);
           expect(ctx.stdout).to.not.contain('Region');
           expect(ctx.stdout.match(/true/g)).to.have.length(1);
-          expect(ctx.stdout).to.match(/project1.*true/);
+          expect(ctx.stdout).to.match(/profile1.*true/);
           expect(ctx.stderr).to.equal('');
         });
 
       test
         .do(ctx => {
           ctx.userConfig = new ConfigData();
-          ctx.userConfig.addProject('default', constants.FAKE_ACCOUNT_SID, 'dev');
+          ctx.userConfig.addProfile('default', constants.FAKE_ACCOUNT_SID, 'dev');
         })
         .twilioCliEnv(Config)
         .stdout()
         .stderr()
-        .twilioCommand(ProjectsList, [])
-        .it('runs projects:list with 1 regional project', ctx => {
+        .twilioCommand(ProfilesList, [])
+        .it('runs profiles:list with 1 regional profile', ctx => {
           expect(ctx.stdout).to.contain('default');
           expect(ctx.stdout).to.contain(constants.FAKE_ACCOUNT_SID);
           expect(ctx.stdout).to.contain('dev');